Create Smallest_diff.cpp
Write a program in SmallestDiff/main.cpp that reads trios of integers from the standard input stream until there are no more values. For each trio, print the smallest difference between the three. For example, if the three numbers are 1, 10, and 100, the smallest difference is 9 (10 - 1), as (100 - 1) and (100 - 10) are both larger than 9.